full card and deck crud functionality

+++ vue/src/services/FlashCardService.js
@@ -25,6 +25,7 @@ listCards() {
  },
 
  deleteDeck(deckID) {
+   if(confirm('Are you sure you want to delete this entire Deck?'))
     return http.delete(`/decks/${deckID}`)
  },
